move roster creator into utility
