+++ platform-sdk/swirlds-platform-core/src/test/java/com/swirlds/platform/util/AddressBookNetworkUtilsTests.java
@@ -16,15 +16,22 @@
 
 package com.swirlds.platform.util;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import com.hedera.hapi.node.state.roster.Roster;
+import com.swirlds.common.platform.NodeId;
 import com.swirlds.common.test.fixtures.Randotron;
+import com.swirlds.platform.builder.PlatformBuilder;
 import com.swirlds.platform.network.Network;
 import com.swirlds.platform.state.address.AddressBookNetworkUtils;
+import com.swirlds.platform.system.BasicSoftwareVersion;
 import com.swirlds.platform.system.address.Address;
 import com.swirlds.platform.system.address.AddressBook;
+import com.swirlds.platform.system.address.AddressBookUtils;
 import com.swirlds.platform.test.fixtures.addressbook.RandomAddressBookBuilder;
 import java.net.Inet4Address;
 import java.net.InetAddress;
@@ -75,5 +82,54 @@ class AddressBookNetworkUtilsTests {
 
         final Address badLocalAddress = address.copySetHostnameInternal("500.8.8");
         assertThrows(IllegalStateException.class, () -> AddressBookNetworkUtils.isLocal(badLocalAddress));
+    }
+
+    @Test
+    void testCreateRosterFromNonEmptyAddressBook() {
+        final PlatformBuilder platformBuilder = PlatformBuilder.create(
+                "name",
+                "swirldName",
+                new BasicSoftwareVersion(1),
+                () -> null,
+                (inputStream, path) -> null,
+                new NodeId(0));
+
+        final Address address1 = new Address(new NodeId(1), "", "", 10, null, 77, null, 88, null, null, "");
+        final Address address2 = new Address(new NodeId(2), "", "", 10, null, 77, null, 88, null, null, "");
+        final AddressBook addressBook = new AddressBook();
+        addressBook.add(address1);
+        addressBook.add(address2);
+        platformBuilder.withBootstrapAddressBook(addressBook);
+        final Roster roster = AddressBookUtils.createRoster(addressBook);
+
+        assertNotNull(roster);
+        assertEquals(2, roster.rosters().size());
+        assertEquals(1L, roster.rosters().getFirst().nodeId());
+        assertEquals(2L, roster.rosters().getLast().nodeId());
+    }
+
+    @Test
+    void testCreateRosterFromNullAddressBook() {
+        assertThrows(
+                NullPointerException.class,
+                () -> AddressBookUtils.createRoster(null),
+                "Illegal attempt to create a Roster from a null AddressBook");
+    }
+
+    @Test
+    void testCreateRosterFromEmptyAddressBook() {
+        final PlatformBuilder platformBuilder = PlatformBuilder.create(
+                "name",
+                "swirldName",
+                new BasicSoftwareVersion(1),
+                () -> null,
+                (inputStream, path) -> null,
+                new NodeId(0));
+        final AddressBook addressBook = new AddressBook();
+        platformBuilder.withBootstrapAddressBook(addressBook);
+        final Roster roster = AddressBookUtils.createRoster(addressBook);
+
+        assertNotNull(roster);
+        assertTrue(roster.rosters().isEmpty());
     }
 }
